+++ apps/bridge-dapp/src/containers/TxInfoContainer/index.ts
@@ -1,0 +1,3 @@
+import TxInfoContainer from './TxInfoContainer';
+
+export default TxInfoContainer;
